add locale picker to user setting
