+++ src/components/index.ts
@@ -88,12 +88,9 @@ export { TranslationProgress } from './TranslationProgress';
 export { TranslationSaveFailed } from './TranslationSaveFailed';
 export { TranslationUser } from './TranslationUser';
 export { TypeRadioGroup } from './TypeRadioGroup';
-export { UserEditForm } from './UserEditForm';
 export { UserEmailEditForm } from './UserEmailEditForm';
 export { UserInvitationList } from './UserInvitationList';
 export { UserPasswordEditForm } from './UserPasswordEditForm';
-export { UserSettingBase } from './UserSettingBase';
-export { UserSettingSafe } from './UserSettingSafe';
 export { VCodeInput } from './VCodeInput';
 export { AdminSiteSetting } from './AdminSiteSetting';
 export { AdminVCodeList } from './AdminVCodeList';
