+++ src/components/setting/UserBasicSettings.tsx
@@ -1,10 +1,12 @@
 import { css } from '@emotion/core';
 import React from 'react';
 import { useIntl } from 'react-intl';
-import { Content, ContentItem, ContentTitle, UserEditForm } from '.';
-import { FC } from '../interfaces';
-import style from '../style';
-import { AvatarUpload } from './AvatarUpload';
+import { Content, ContentItem, ContentTitle } from '..';
+import { FC } from '../../interfaces';
+import style from '../../style';
+import { AvatarUpload } from '../AvatarUpload';
+import { UserEditForm } from './UserEditForm';
+import { LocalePicker } from './LocalePicker';
 
 /** 用户基础设置的属性接口 */
 interface UserSettingBaseProps {
@@ -13,7 +15,7 @@ interface UserSettingBaseProps {
 /**
  * 用户基础设置
  */
-export const UserSettingBase: FC<UserSettingBaseProps> = ({ className }) => {
+export const UserBasicSettings: FC<UserSettingBaseProps> = ({ className }) => {
   const { formatMessage } = useIntl(); // i18n
 
   return (
@@ -26,6 +28,7 @@ export const UserSettingBase: FC<UserSettingBaseProps> = ({ className }) => {
       `}
     >
       <Content>
+        <LocalePicker />
         <ContentTitle>{formatMessage({ id: 'user.info' })}</ContentTitle>
         <div style={{ marginBottom: '24px' }}>
           <AvatarUpload type="user" />
